fix: 실제 요청 데이터에 따른 타입 변경

+++ src/main/java/com/ssafy/sandbox/paging/dto/RequestCursor.java
@@ -1,4 +1,4 @@
 package com.ssafy.sandbox.paging.dto;
 
-public record RequestCursor(int size, int cursorId) {
+public record RequestCursor(int size, Long cursorId) {
 }
